letsgo

+++ KG_logger/can_logger.py
@@ -12,22 +12,30 @@ import threading
 
 app = typer.Typer()
 
-
-def upload_to_dropbox_async(log_file, dropbox_token, dropbox_path="/"):
-    def upload_file():
+def upload_to_dropbox_async(log_files, dropbox_token, dropbox_path="/", stop_event=None):
+    def upload_files():
         dbx = dropbox.Dropbox(dropbox_token)
-        with open(log_file, "rb") as f:
-            file_data = f.read()
-            try:
-                dbx.files_upload(file_data, dropbox_path + log_file.name, mode=WriteMode('overwrite'))
-                os.remove(log_file)  # Удаляем файл после успешной загрузки
-                logging.info(f"Uploaded {log_file} to Dropbox")
-            except dropbox.exceptions.ApiError as e:
-                logging.error(f"Failed to upload {log_file} to Dropbox: {e}")
+        for log_file in log_files:
+            if stop_event and stop_event.is_set():
+                return
+            with open(log_file, "rb") as f:
+                file_data = f.read()
+                try:
+                    dbx.files_upload(file_data, dropbox_path + log_file.name, mode=WriteMode('overwrite'))
+                    os.remove(log_file)  # Удаляем файл после успешной загрузки
+                    logging.info(f"Uploaded {log_file} to Dropbox")
+                except dropbox.exceptions.ApiError as e:
+                    logging.error(f"Failed to upload {log_file} to Dropbox: {e}")
+                    return  # Прекращаем попытки загрузки при ошибке
+    threading.Thread(target=upload_files).start()
 
-    upload_thread = threading.Thread(target=upload_file)
-    upload_thread.start()
-
+def check_internet_connection():
+    try:
+        dbx = dropbox.Dropbox("fake_token")  # Проверка без токена
+        dbx.users_get_current_account()  # Запрос к API Dropbox для проверки соединения
+        return True
+    except:
+        return False
 
 def read_temperatures(sensors, sensor_count, interval, stop_event, temperatures):
     while not stop_event.is_set():
@@ -42,15 +50,15 @@ def read_temperatures(sensors, sensor_count, interval, stop_event, temperatures)
             temperatures[i] = temperature
         time.sleep(interval)
 
-
 @app.command()
 def log_can_data(interface: str = typer.Argument("can0", help="CAN interface, e.g., can0"),
                  log_dir: str = typer.Argument("./logs", help="Directory to save log files"),
-                 log_duration: int = typer.Argument(10, help="Log duration in seconds"),
+                 log_duration: int = typer.Argument(10, help="Log duration in minutes"),
                  dropbox_token: str = typer.Option(..., help="Dropbox API OAuth2 token"),
                  dropbox_path: str = typer.Option("/", help="Dropbox path to upload files"),
                  log_name: str = typer.Option("can_log", help='Optional base name for the log file')
                  ):
+
     Path(log_dir).mkdir(parents=True, exist_ok=True)
     logging.basicConfig(level=logging.DEBUG)
     logger = logging.getLogger('CAN_Logger')
@@ -70,9 +78,7 @@ def log_can_data(interface: str = typer.Argument("can0", help="CAN interface, e.
 
         file_handler = logging.FileHandler(log_file, mode='w')
         logger.addHandler(file_handler)
-        # Добавляем заголовок CSV
-        logger.info(
-            "Timestamp,ID,Ext,RTR,Dir,Bus,Len,Data,Temperature_1,Temperature_2,Temperature_3,Temperature_4,Temperature_5,Temperature_6")
+        logger.info("Timestamp,ID,Ext,RTR,Dir,Bus,Len,Data,Temperature_1,Temperature_2,Temperature_3,Temperature_4,Temperature_5,Temperature_6")
         return log_file
 
     log_file = rotate_log_file()
@@ -82,7 +88,6 @@ def log_can_data(interface: str = typer.Argument("can0", help="CAN interface, e.
     temperatures = ["Unavailable"] * 6
     stop_event = threading.Event()
 
-    # Запускаем поток для считывания данных с датчиков
     temp_thread = threading.Thread(target=read_temperatures, args=(sensors, sensor_count, 2, stop_event, temperatures))
     temp_thread.start()
 
@@ -96,22 +101,25 @@ def log_can_data(interface: str = typer.Argument("can0", help="CAN interface, e.
             log_entry = f"{datetime.now().isoformat()},{hex(msg.arbitration_id)},{msg.is_extended_id},{msg.is_remote_frame},{msg.is_error_frame},{msg.channel},{msg.dlc},{data_str},{','.join(map(str, temperatures))}"
             logger.info(log_entry)
 
-            # Проверяем, прошло ли заданное время для смены файла
             if datetime.now() - log_start_time >= timedelta(seconds=log_duration):
-                upload_to_dropbox_async(log_file, dropbox_token, dropbox_path)
                 log_file = rotate_log_file()
+
+            if check_internet_connection():
+                log_files = list(Path(log_dir).glob(f"{log_name}_*.csv"))
+                upload_to_dropbox_async(log_files, dropbox_token, dropbox_path, stop_event)
 
     except (OSError, can.CanError) as e:
         logger.error(f"Error with CAN interface: {e}")
     except KeyboardInterrupt:
-        logger.warning("KeyboardInterrupt received, saving and uploading log file.")
-        upload_to_dropbox_async(log_file, dropbox_token, dropbox_path)
-    finally:
+        logger.warning("KeyboardInterrupt received, saving and uploading log files.")
         stop_event.set()  # Останавливаем поток считывания температур
         temp_thread.join()
+
+        log_files = list(Path(log_dir).glob(f"{log_name}_*.csv"))
+        upload_to_dropbox_async(log_files, dropbox_token, dropbox_path, stop_event)
+    finally:
         if logger.handlers:
             logger.handlers[0].close()
-
 
 if __name__ == "__main__":
     app()
